rf: 调整代码结构
