+++ src/types/mns-types.ts
@@ -1,0 +1,8 @@
+export type MNSOptions = {
+  accountId: string;
+  regionId: string;
+  accessKeyId: string;
+  accessKeySecret: string;
+};
+
+export type MNSQueueOptions = { queueName: string } & MNSOptions;
